segundo commit

+++ app/src/main/java/com/turingalan/pokemon/data/repository/PokemonInMemoryRepository.kt
@@ -60,9 +60,9 @@ class PokemonInMemoryRepository @Inject constructor(): PokemonRepository {
         )
     }
 
-    override fun readOne(id: Int): Pokemon? {
+    override fun readOne(id: Long): Pokemon? {
         val pokemon = pokemonList().firstOrNull() {
-            p -> p.id == id
+            p -> p.id.toInt().toLong() == id
         }
         return pokemon
     }
